\name{anno.methyl}
\alias{anno.methyl}
\docType{data}
\title{
Methylation to gene annotation file 
}
\description{}
\usage{data("anno.methyl")}
\format{
  The format is:
 chr [1:174, 1:2] "YWHAE" "EIF4EBP1" "TP53BP1" "ACACA" "ACACB" "AKT1" "AKT2" "AKT3" "PRKAA1" "AR" "ASNS" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "gene" "Methylation"
}

