\name{anno.miRNA}
\alias{anno.miRNA}
\docType{data}
\title{
miRNA to Gene annotation file}
\description{}
\usage{data("anno.miRNA")}
\format{
  The format is:
 chr [1:502058, 1:7] "hsa-miR-647" "hsa-miR-130a" "hsa-miR-423-3p" "hsa-miR-423-5p" "hsa-miR-130b" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:7] "name" "target" "chrom" "start" ...
}
