\name{Methylation}
\alias{Methylation}
\docType{data}
\title{
Methylation data for TCGA KIRC patients 
}
\description{}
\usage{data("Methylation")}
\format{
  The format is:
List of 2
 $ Data: num [1:324, 1:454] 0.0669 0.0253 0.1005 0.9372 0.9612 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:454] "TCGA-A3-3322-01" "TCGA-A3-3346-01" "TCGA-A3-3365-01" "TCGA-A3-3380-01" ...
 $ Des : chr [1:324, 1:4] "cg21137823" "cg25007680" "cg19560758" "cg04508649" ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:4] "REF" "GeneSymbol" "ChromosomeID" "CoordinateID"
}
