\name{miRNA}
\alias{miRNA}
\docType{data}
\title{
miRNA expression data for TCGA KIRC patients}
\description{}
\usage{data("miRNA")}
\format{
  The format is:
List of 2
 $ Data: num [1:296, 1:454] 2.38e+01 1.68e-03 -4.62e-03 4.99e-04 2.92e+02 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:454] "TCGA-A3-3322-01" "TCGA-A3-3346-01" "TCGA-A3-3365-01" "TCGA-A3-3380-01" ...
 $ Des : chr [1:296] "hsa-mir-185" "hsa-mir-586" "hsa-mir-519d" "hsa-mir-520h" ...
}
