\name{mRNA}
\alias{mRNA}
\docType{data}
\title{
mRNA expression data for TCGA KIRC patients}
\description{}
\usage{data("mRNA")}
\format{
  The format is:
List of 2
 $ Data: num [1:174, 1:454] 8150 432 989 1016 1184 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:454] "TCGA-A3-3322-01" "TCGA-A3-3346-01" "TCGA-A3-3365-01" "TCGA-A3-3380-01" ...
 $ Des : chr [1:174, 1:2] "YWHAE" "EIF4EBP1" "TP53BP1" "ACACA" ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:2] "GeneSymbol" "EntrezID"
}
