\name{ppi}
\alias{ppi}
\docType{data}
\title{
Pathway information and Protein-Protein interactions from String DB V10
}
\description{}
\usage{data("ppi")}
\format{
  The format is:
List of 2
 $ pathwaydat: chr [1:87, 1:3] "Apoptosis" "Apoptosis" "Apoptosis" "Apoptosis" ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:3] "Pathway_v20150123" "Protein" "gene"
 $ ppi       :List of 12
  ..$ Apoptosis                 : chr [1:70, 1:3] "BIRC2" "BIRC2" "BIRC2" "BIRC2" ...
  ..$ Cell cycle                : chr [1:42, 1:3] "CDKN1B" "CDKN1B" "CDKN1B" "CDKN1B" ...
  ..$ DNA damage response       : chr [1:90, 1:3] "XRCC1" "XRCC1" "XRCC1" "XRCC1" ...
  ..$ EMT                       : chr [1:30, 1:3] "SERPINE1" "SERPINE1" "SERPINE1" "SERPINE1" ...
  ..$ Hormone receptor          : chr [1:6, 1:3] "ESR1" "ESR1" "PGR" "PGR" ...
  ..$ Hormone signaling (Breast): chr [1:2, 1:3] "BCL2" "GATA3" "GATA3" "BCL2" ...
  ..$ PI3K/AKT                  : chr [1:74, 1:3] "TSC2" "TSC2" "TSC2" "TSC2" ...
  ..$ RAS/MAPK                  : chr [1:84, 1:3] "MAPK1" "MAPK1" "MAPK1" "MAPK1" ...
  ..$ RTK                       : chr [1:20, 1:3] "ERBB3" "ERBB3" "ERBB3" "ERBB3" ...
  ..$ TSC/mTOR                  : chr [1:18, 1:3] "RPS6KB1" "RPS6KB1" "RPS6KB1" "RPS6KB1" ...
  ..$ Breast reactive           : chr [1:28, 1:3] "GAPDH" "GAPDH" "GAPDH" "GAPDH" ...
  ..$ Core reactive             : chr [1:10, 1:3] "CDH1" "CDH1" "CDH1" "CAV1" ...
}
