\name{rppadat}
\alias{rppadat}
\docType{data}
\title{
RPPA data for TCGA KIRC patients}
\description{}
\usage{data("rppadat")}
\format{
  The format is:
List of 12
 $ : num [1:469, 1:9] 2.11 2.11 2.21 1.51 1.53 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:9] "BAK1" "BAX" "BID" "BCL2L11" ...
 $ : num [1:469, 1:6] 3.06 2.62 1.35 3.02 3.92 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:6] "CAV1" "MYH11" "RAB11A, RAB11B" "CTNNB1" ...
 $ : num [1:469, 1:7] 0.139 0.701 1 0.396 0.551 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:7] "CDK1" "CCNB1" "CCNE1" "CCNE2" ...
 $ : num [1:469, 1:5] 3.06 2.62 1.35 3.02 3.92 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:5] "CAV1" "CTNNB1" "RBM15" "CDH1" ...
 $ : num [1:469, 1:10] 1.221 0.882 1.153 1.007 1.103 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:10] "TP53BP1" "ATM" "CHEK1" "CHEK2" ...
 $ : num [1:469, 1:7] -0.3783 -0.2273 0.1743 -0.1085 0.0971 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:7] "FN1" "CDH2" "COL6A1" "CLDN7" ...
 $ : num [1:469, 1:7] -0.402 0.385 1.925 0.277 0.523 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:7] "AKT1, AKT2, AKT3" "GSK3A, GSK3B" "CDKN1B" "AKT1S1" ...
 $ : num [1:469, 1:9] -0.806 -1.097 -1.131 -1.162 -1.211 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:9] "ARAF" "JUN" "RAF1" "MAPK8" ...
 $ : num [1:469, 1:5] 0.525 0.456 -0.098 0.344 0.612 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:5] "EGFR" "ERBB2" "ERBB3" "SHC1" ...
 $ : num [1:469, 1:5] 0.5228 0.3334 0.3058 -0.0702 -0.0154 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:5] "EIF4EBP1" "RPS6KB1" "MTOR" "RPS6" ...
 $ : num [1:469, 1:3] -0.909 -0.909 -0.967 -0.871 -0.714 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:3] "ESR1" "PGR" "AR"
 $ : num [1:469, 1:3] -1.765 -1.138 -1.691 -0.974 -1.171 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:3] "INPP4B" "GATA3" "BCL2"
}
